refactor(console): merge OSS survey reporting into submit flow
